+++ audiospeech_test.go
@@ -27,9 +27,11 @@ func TestAudioSpeechNewWithOptionalParams(t *testing.T) {
 		option.WithAPIKey("My API Key"),
 	)
 	resp, err := client.Audio.Speech.New(context.TODO(), openai.AudioSpeechNewParams{
-		Input:          "input",
-		Model:          openai.SpeechModelTTS1,
-		Voice:          openai.AudioSpeechNewParamsVoiceAsh,
+		Input: "input",
+		Model: openai.SpeechModelTTS1,
+		Voice: openai.AudioSpeechNewParamsVoiceUnion{
+			OfString: openai.String("string"),
+		},
 		Instructions:   openai.String("instructions"),
 		ResponseFormat: openai.AudioSpeechNewParamsResponseFormatMP3,
 		Speed:          openai.Float(0.25),