+++ chatcompletion_test.go
@@ -38,7 +38,9 @@ func TestChatCompletionNewWithOptionalParams(t *testing.T) {
 		Model: shared.ChatModelGPT5_4,
 		Audio: openai.ChatCompletionAudioParam{
 			Format: openai.ChatCompletionAudioParamFormatWAV,
-			Voice:  openai.ChatCompletionAudioParamVoiceAsh,
+			Voice: openai.ChatCompletionAudioParamVoiceUnion{
+				OfString: openai.String("string"),
+			},
 		},
 		FrequencyPenalty: openai.Float(-2),
 		FunctionCall: openai.ChatCompletionNewParamsFunctionCallUnion{